update .gitignore
the therecipe/qt project offers the possibility to publish your projekt
via docker for sailfish. While building the app there are temporary files rcc*
are created and the output is moved to deploy. these should not be present in
the repo, therefore the gitignore was extended.

also the vendor entry was removed to store the dependencies and to guarantee a
successful build.